+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 | --- | --- | --- |
 | - | ci: :chrone: upgrade vite to v4.2.0 | ([e272d42](https://github.com/chenshunliang/dashboard/commit/e272d42e3664b2cd2522c5f568f99e04a303385b)) |
 | - | ci: :chrone: update vite config | ([b1e9894](https://github.com/chenshunliang/dashboard/commit/b1e9894e7b52a6ebddacabe24bdeead589f2d2c2)) |
+| - | ci: :chrone: update vie chunk cplit config | ([a044d42](https://github.com/chenshunliang/dashboard/commit/a044d427e3747a54d57576851c40b3ac61a06673)) |
 
 # 1.23.6 (2023-3-10)