+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.23.7 (2023-3-20)
+# 1.23.7 (2023-3-22)
 
 ### Bug Fixes
 
@@ -6,6 +6,7 @@
 | --- | --- | --- |
 | - | [fix: 🐛 fix json schema render bug and fix list sort bug] | [9e1ee94](https://github.com/chenshunliang/dashboard/commit/9e1ee94f81e82add42583bc637f500249a108ec9) |
 | - | [fix: 🐛 fix json schema init bug, fix app store list data bug] | [f2ff4f2](https://github.com/chenshunliang/dashboard/commit/f2ff4f261febb146156714725d117a521ecf6d4b) |
+| - | [fix: 🐛 fix resource approve api bug] | [5c7199a](https://github.com/chenshunliang/dashboard/commit/5c7199a21e4b6ee45203ea116030519db1fa4c72) |
 
 ### 🔨 reflactor